Fix playback behavior bugs.
First, get rid of repetitive call to handlePlayRequest() in QueueManager
since it is already triggered by callbacks in MediaSessionCallback.
Second, clicking queue item in TvPlaybackFragment only skips to that
item (plays that item) if it is the current media item  AND paused, or if
it is a new media item.

Change-Id: Iecdd153a137e64526a4f3380fd0d08e94a7bc833