Ensure metadata is propagated.
This change works around a peculiarity in the ExoPlayer media session
extension and how metadata is propagated.

By including a download status, MediaMetadataCompat will create an
"extras" bundle, which will then be included when retreiving
its MediaDescriptionCompat (with `.getDescription()`)

This in turn allows the placing of *all* of the metadata from
the MediaMetadataCompat into the MediaDescriptionCompat that
`.getDescription()` returns, which is then passed along to the
media session as one would expect.

This should fix the issue of album art not appearing on the
lockscreen: GitHub issue
[#227](https://github.com/googlesamples/android-UniversalMusicPlayer/issues/227).

Change-Id: I924dfd8caf0166e60b9d405e769d8cfa7f2351dc