WIP to use content:// Uris for album art.
This addresses #235

At the moment, the first time art is retrieved it doesn't appear.
Subsequent loads will show the images.

This is likely related to Glide not observing the Uri for changes. This
might work on other form factors (for example, Automotive OS) but it
also might be necessary to update the MediaItems themselves.

The change is probably going to wait until January 2021 to be resolved.